backend/requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Rate Limiting
Flask-Limiter==3.5.0

# HTTP Requests
requests==2.31.0

# Redis for caching
redis==5.0.1

# OpenAI API
openai==1.3.7

# Data Processing
pandas==2.1.4
numpy==1.25.2

# JSON handling
ujson==5.8.0

# Environment variables
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# Date/Time handling
python-dateutil==2.8.2

# Validation
marshmallow==3.20.1

# Caching utilities
cachetools==5.3.2

# HTTP client with retry
urllib3==2.1.0

# Development dependencies (optional)
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Production server
gunicorn==21.2.0

# Monitoring
psutil==5.9.6

